Clean up component structure
